add defaultValues to ProjectDialogProps
