+++ src/components/organisms/dialog/ProjectDialog/index.tsx
@@ -2,6 +2,7 @@ import { LoadableAction } from '@/contexts/openapi';
 import { ProjectWithoutAutofield, ProjectWithoutAutofieldError } from '@/openapi';
 
 import ProjectDialogComponent from './ProjectDialog';
+import { ProjectFieldValues } from './ProjectDialogForm';
 
 import React from 'react';
 
@@ -9,6 +10,7 @@ export type ProjectDialogProps = {
   readonly open: boolean;
   readonly title: string;
   readonly submitText: string;
+  readonly defaultValues: ProjectFieldValues;
   readonly onSubmit: (project: ProjectWithoutAutofield) => Promise<LoadableAction<ProjectWithoutAutofieldError>>;
   readonly onClose: () => void;
 };
